--- src/main/java/com/noteu/noteu/member/security/handler/MemberAccessDeniedHandler.java
@@ -1,26 +1,0 @@
-package com.noteu.noteu.member.security.handler;
-
-import com.noteu.noteu.member.security.utils.ErrorResponder;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-
-//서버에 요청을 할 때 액세스가 가능한지 권한을 체크후 액세스 할 수 없는 요청을 했을시 동작된다.
-@Slf4j
-@Component
-public class MemberAccessDeniedHandler implements AccessDeniedHandler {
-    @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException, ServletException {
-
-        ErrorResponder.sendErrorResponse(response, HttpStatus.FORBIDDEN);
-        log.warn("Forbidden error happened: {}", accessDeniedException.getMessage());
-    }
-}
