--- src/main/java/com/noteu/noteu/member/security/handler/MemberAuthenticationEntryPoint.java
@@ -1,30 +1,0 @@
-package com.noteu.noteu.member.security.handler;
-
-import com.noteu.noteu.member.security.utils.ErrorResponder;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Slf4j
-@Component
-public class MemberAuthenticationEntryPoint implements AuthenticationEntryPoint { // 인증되지 않은 사용자 처리
-    @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException, ServletException {
-        Exception exception = (Exception) request.getAttribute("exception");
-        ErrorResponder.sendErrorResponse(response, HttpStatus.UNAUTHORIZED);
-
-        logExceptionMessage(authException, exception);
-    }
-
-    private void logExceptionMessage(AuthenticationException authException, Exception exception) {
-        String message = exception != null ? exception.getMessage() : authException.getMessage();
-        log.warn("Unauthorized error happened: {}", message);
-    }
-}
