--- src/main/java/com/noteu/noteu/member/security/handler/MemberAuthenticationSuccessHandler.java
@@ -1,38 +1,0 @@
-package com.noteu.noteu.member.security.handler;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.noteu.noteu.member.dto.MemberDetails;
-import com.noteu.noteu.member.dto.MemberInfo;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-
-import java.io.IOException;
-
-@Slf4j
-public class MemberAuthenticationSuccessHandler implements AuthenticationSuccessHandler {
-
-    @Override
-    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
-        ObjectMapper mapper = new ObjectMapper();
-        MemberDetails memberDetails = (MemberDetails) authentication.getPrincipal();
-        log.info("MemberAuthenticationSuccessHandler auth.getPrincipal : {}", authentication.getPrincipal().toString());
-
-        MemberInfo memberInfo = MemberInfo.builder()
-                .id(memberDetails.getId())
-                .username(memberDetails.getUsername())
-                .authorities(memberDetails.getAuthorities())
-                .build();
-        String memberInfoToJson = mapper.writeValueAsString(memberInfo);
-        response.setContentType("application/json");
-        response.getWriter().write(memberInfoToJson);
-        log.info("login authentication success!!!");
-
-        String authorization = response.getHeader("Authorization");
-        log.info("successHandler 토큰 값: {}", authorization);
-    }
-}
