+++ src/main/resources/templates/fragments/member/login.html
@@ -1,0 +1,43 @@
+<!DOCTYPE html>
+<html lang="en" data-layout-mode="detached" data-topbar-color="dark" data-menu-color="light" data-sidenav-user="true" xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
+<th:block th:replace="fragments/common/header :: headerFragment"></th:block>
+<head>
+    <!-- 폰트어썸 아이콘 -->
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css">
+</head>
+<body>
+<div class="wrapper">
+    <th:block th:replace="fragments/common/topbar :: topbarFragment"></th:block>
+    <th:block th:replace="fragments/common/sidebar1 :: sidebar1Fragment"></th:block>
+    <div class="content-page">
+        <div class="content">
+            <h3 class="my-3">로그인</h3>
+            <form th:action="@{/auth/login}" method="post">
+                <div th:if="${param.error}">
+                    <div class="alert alert-danger">
+                        ID 또는 비밀번호를 확인해 주세요.
+                    </div>
+                </div>
+            <div class="mb-2">
+                <label for="username">아이디</label>
+                <input type="text" name="username" id="username" class="form-control"><br/>
+            </div>
+            <div class="mb-2">
+                <label for="password">비밀번호</label>
+                <input type="password" name="password" id="password" class="form-control"><br/>
+            </div>
+            <input type="submit" id="login_btn" value="login" class="btn btn-primary float-end">
+            </form>
+            <div class="flex flex-col w-11/12 gap-2">
+                <a href="/oauth2/authorization/kakao" class="btn btn-warning gap-2 bg-yellow-300">
+                    <i class="fa-solid fa-comment"></i>
+                    <span>카카오 로그인</span>
+                </a>
+            </div>
+        </div>
+        <th:block th:replace="fragments/common/footer :: footerFragment"></th:block>
+    </div>
+    <th:block th:replace="fragments/common/script :: scriptFragment"></th:block>
+</div>
+</body>
+</html>
