--- src/main/resources/templates/fragments/member/sign_in.html
@@ -1,77 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" data-layout-mode="detached" data-topbar-color="dark" data-menu-color="light" data-sidenav-user="true" xmlns:th="http://www.thymeleaf.org" xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
-<th:block th:replace="fragments/common/header :: headerFragment"></th:block>
-<head>
-    <!-- 폰트어썸 아이콘 -->
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css">
-
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.4/jquery.min.js"></script>
-    <script src="/webjars/axios/0.17.1/dist/axios.min.js"></script>
-    <script type="text/javascript">
-        window.onload = function() {
-            $("#login_btn").click(function () {
-                var username = $("#username").val();
-                var password = $("#password").val();
-                if (username === '') {
-                    alert("아이디를 입력하세요.");
-                    return;
-                }
-                if (password === '') {
-                    alert("패스워드를 입력하세요.");
-                    return;
-                }
-                axios.post('/members/sign-in', {
-                    username: username,
-                    password: password
-                })
-                    .then(response => {
-                        console.log(response);
-                        console.log(response.headers.authorization);
-                        if (response.data === false) {
-                            alert("로그인 실패");
-                        } else {
-                            let token = document.cookie.split("=")[1];
-                            axios.get("/subjects", {
-                                headers: {
-                                    Authorization: `${token}`
-                                },
-                                url: "/subjects"
-                            })
-                        }
-                    })
-                    .catch(response => {
-                        alert("POST 요청 실패");
-                    });
-            });
-        }
-    </script>
-</head>
-<body>
-<div class="wrapper">
-    <th:block th:replace="fragments/common/topbar :: topbarFragment"></th:block>
-    <th:block th:replace="fragments/common/sidebar1 :: sidebar1Fragment"></th:block>
-    <div class="content-page">
-        <div class="content">
-            <h3 class="my-3">로그인</h3>
-            <div class="mb-2">
-                <label for="username">아이디</label>
-                <input type="text" name="username" id="username" class="form-control"><br/>
-            </div>
-            <div class="mb-2">
-                <label for="password">비밀번호</label>
-                <input type="password" name="password" id="password" class="form-control"><br/>
-            </div>
-            <div class="flex flex-col w-11/12 gap-2">
-                <a href="/oauth2/authorization/kakao" class="btn btn-warning gap-2 bg-yellow-300">
-                    <i class="fa-solid fa-comment"></i>
-                    <span>카카오 로그인</span>
-                </a>
-            </div>
-            <input type="button" id="login_btn" value="login" class="btn btn-primary float-end">
-        </div>
-        <th:block th:replace="fragments/common/footer :: footerFragment"></th:block>
-    </div>
-    <th:block th:replace="fragments/common/script :: scriptFragment"></th:block>
-</div>
-</body>
-</html>
